Merged PR 4928: convert the geometry data to FME workbench friendly string within the framework
Currently the cleansed tables in hydra seem to have geometry data that is unreadable in FME Workbench. The double apostrophes around the coordinates need to be cleaned by end user in order to make it usable. The request is to make that transformation happen within the framework. More details in the jira ticket:

https://elogin.ads.swc/jira/browse/DAF-9542

This PR resolves that problem by declaring a udf and using that udf for geometry columns.